shift pop tests to dependency tests
